Moving `IndexHelper` tests to corresponding suite.
